+++ source/LibraryLink/Himalaya_LibraryLink.m
@@ -80,7 +80,7 @@ Output: A list of replacement rules with
  - Mh2ShiftDRbarPrimeToMDRPrime: list with {0L, 1L, 2L, 3L} shifts from DR' -> MDR' scheme
  - Mh2ShiftDRbarPrimeToH3m: list with {0L, 1L, 2L, 3L} shifts from DR' -> H3m scheme
  - expansionUncertainty: uncertainty from mass expansion for 0-, 1-, 2- and 3-loop correction
- - Mh2EFT: list with {0L, 1L, 2L, 3L} squared light CP-even Higgs mass contribution in the decoupling limit (DR' scheme) (only at-enhanced contributions)
+ - Mh2EFTAt: list with {0L, 1L, 2L, 3L} squared light CP-even Higgs mass contribution in the decoupling limit (DR' scheme) (only at-enhanced contributions)
  - Mh2FO: list with {0L, 1L, 2L, 3L} squared light CP-even Higgs mass contribution (DR' scheme)
  - Mh2FOAt: list with {0L, 1L, 2L, 3L} squared light CP-even Higgs mass contribution (DR' scheme) (only at-enhanced contributions)
  - lambda: {0L, 1L, 2L, 3L} correction to lambda, in DR' scheme, w/ full logarithms
@@ -153,7 +153,7 @@ output = HimalayaCalculateDMh3L[
   Mh2ShiftDRbarPrimeToMDRPrime,
   Mh2ShiftDRbarPrimeToH3m,
   expansionUncertainty,
-  Mh2EFT, Mh2FO, Mh2FOAt,
+  Mh2EFTAt, Mh2FO, Mh2FOAt,
   lambda, lambdaUncertainty, lambdaShiftDRbarPrimeToMSbar };
 
 Himalaya::nonum = "Error: `1` is not a numeric input value!";